Move wirelesstag base entity to separate module
