+++ homeassistant/components/wirelesstag/binary_sensor.py
@@ -15,8 +15,8 @@ from homeassistant.helpers.dispatcher import async_dispatcher_connect
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.typing import ConfigType, DiscoveryInfoType
 
-from . import WirelessTagBaseSensor
 from .const import DOMAIN, SIGNAL_BINARY_EVENT_UPDATE
+from .entity import WirelessTagBaseSensor
 from .util import async_migrate_unique_id
 
 # On means in range, Off means out of range
