+++ homeassistant/components/wirelesstag/switch.py
@@ -17,8 +17,8 @@ import homeassistant.helpers.config_validation as cv
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.typing import ConfigType, DiscoveryInfoType
 
-from . import WirelessTagBaseSensor
 from .const import DOMAIN
+from .entity import WirelessTagBaseSensor
 from .util import async_migrate_unique_id
 
 SWITCH_TYPES: tuple[SwitchEntityDescription, ...] = (
